web_app/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
boto3==1.42.44
pydantic==2.9.2
requests==2.31.0
python-dotenv==1.0.1
PyJWT==2.8.0
passlib==1.7.4
bcrypt==3.2.2
pinterest-api-sdk
aioboto3==12.3.0